requirements_dev.txt
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
fakeredis==2.19.0

# Code Quality
black==23.12.0
ruff==0.1.8
mypy==1.7.1
pylint==3.0.3

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Debugging
ipython==8.18.1
ipdb==0.13.13

# Type Stubs
types-redis==4.6.0.11
sqlalchemy-stubs==0.4

# Pre-commit hooks
pre-commit==3.5.0
bandit==1.7.5
